+++ react-auth-form/src/components/LoginInput/style.module.css
@@ -18,3 +18,18 @@
   border-color: #007bff;
   box-shadow: 0 0 4px rgba(0, 123, 255, 0.25);
 }
+
+.errorMessage {
+  color: red;
+  font-size: 0.75rem;
+}
+
+.submitButton {
+  padding: 0.8rem;
+  background-color: #007bff;
+
+  color: white;
+  border: none;
+  border-radius: 0.5rem;
+  font-size: 1rem;
+}
