--- react-auth-form/src/hooks/useLogin.tsx
@@ -1,23 +1,0 @@
-import { useState } from "react";
-
-const useLogin = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  const handleEmail = (email: string) => {
-    setEmail(email);
-  };
-
-  const handlePassword = (password: string) => {
-    setPassword(password);
-  };
-
-  return {
-    email,
-    handleEmail,
-    password,
-    handlePassword,
-  };
-};
-
-export default useLogin;
